+++ src/nsidc/metgen/metgen.py
@@ -233,6 +233,7 @@ class Granule:
     collection: Maybe[Collection] = Maybe.empty
     data_filenames: set[str] = dataclasses.field(default_factory=set)
     browse_filenames: set[str] = dataclasses.field(default_factory=set)
+    premet_filename: Maybe[str] = Maybe.empty
     ummg_filename: Maybe[str] = Maybe.empty
     submission_time: Maybe[str] = Maybe.empty
     uuid: Maybe[str] = Maybe.empty
@@ -300,9 +301,12 @@ def process(configuration: config.Config) -> None:
             name,
             data_filenames=data_files,
             browse_filenames=browse_files,
+            premet_filename=premet_file,
             data_reader=data_reader(configuration.auth_id, data_files),
         )
-        for name, data_files, browse_files in grouped_granule_files(configuration)
+        for name, data_files, browse_files, premet_file in grouped_granule_files(
+            configuration
+        )
     ]
     granules = take(configuration.number, candidate_granules)
     results = [pipeline(g) for g in granules]
@@ -517,10 +521,10 @@ def collection_from_cmr(environment: str, auth_id: str, version: int):
 
 def grouped_granule_files(configuration: config.Config) -> list[tuple]:
     """
-    Identify data file(s) and browse file(s) related to each granule.
+    Identify file(s) related to each granule.
     """
     file_list = [p for p in Path(configuration.data_dir).glob("*")]
-    granule_keys = find_granule_keys(configuration, file_list)
+    premet_file_list = premet_files(configuration)
 
     return [
         granule_tuple(
@@ -528,12 +532,29 @@ def grouped_granule_files(configuration: config.Config) -> list[tuple]:
             configuration.granule_regex or f"({granule_key})",
             configuration.browse_regex,
             file_list,
+            premet_file_list,
         )
-        for granule_key in granule_keys
+        for granule_key in granule_keys(configuration, file_list)
     ]
 
 
-def find_granule_keys(configuration: config.Config, file_list: list[Path]) -> set[str]:
+def premet_files(configuration: config.Config) -> list[Path]:
+    if configuration.premet_dir:
+        premets = [
+            p
+            for p in Path(configuration.premet_dir).glob(f"*{constants.PREMET_SUFFIX}")
+        ]
+        if not premets:
+            raise Exception(
+                f"Premet directory {configuration.premet_dir} is empty or unreadable."
+            )
+
+        return premets
+
+    return None
+
+
+def granule_keys(configuration: config.Config, file_list: list[Path]) -> set[str]:
     if configuration.granule_regex:
         return granule_keys_from_regex(configuration.granule_regex, file_list)
     else:
@@ -565,7 +586,11 @@ def granule_keys_from_filename(browse_regex, file_list):
 
 
 def granule_tuple(
-    granule_key: str, granule_regex: str, browse_regex: str, file_list: list
+    granule_key: str,
+    granule_regex: str,
+    browse_regex: str,
+    file_list: list,
+    premet_list: list,
 ) -> tuple:
     """
     Important! granule_regex argument must include a captured match group
@@ -576,6 +601,7 @@ def granule_tuple(
           otherwise the common name elements of all files related to a granule.
         - A set of one or more full paths to data file(s)
         - A set of zero or more full paths to associated browse file(s)
+        - Path to an associated premet file (may be None or empty string)
     """
     browse_file_paths = {
         str(file)
@@ -587,10 +613,22 @@ def granule_tuple(
         str(file) for file in file_list if re.search(granule_key, file.name)
     } - browse_file_paths
 
+    if premet_list is None:
+        premet_file = None
+    else:
+        premet_matches = [
+            str(file) for file in premet_list if re.search(granule_key, file.name)
+        ]
+        if not premet_matches:
+            premet_file = ""
+        else:
+            premet_file = first(premet_matches)
+
     return (
         derived_granule_name(granule_regex, data_file_paths),
         data_file_paths,
         browse_file_paths,
+        premet_file,
     )
 
 
@@ -629,9 +667,13 @@ def prepare_granule(_: config.Config, granule: Granule) -> Granule:
     )
 
 
+def derived_ummg_filename(ummg_path: Path, granule_id: str) -> Path:
+    return ummg_path.joinpath(granule_id + ".json")
+
+
 def find_existing_ummg(configuration: config.Config, granule: Granule) -> Granule:
-    ummg_filename = configuration.ummg_path().joinpath(
-        granule.producer_granule_id + ".json"
+    ummg_filename = derived_ummg_filename(
+        configuration.ummg_path(), granule.producer_granule_id
     )
 
     if ummg_filename.exists():
@@ -648,8 +690,8 @@ def create_ummg(configuration: config.Config, granule: Granule) -> Granule:
     if granule.ummg_filename != Maybe.empty and not configuration.overwrite_ummg:
         return granule
 
-    ummg_file_path = configuration.ummg_path().joinpath(
-        granule.producer_granule_id + ".json"
+    ummg_file_path = derived_ummg_filename(
+        configuration.ummg_path(), granule.producer_granule_id
     )
 
     # Populated metadata_details dict looks like:
@@ -665,7 +707,9 @@ def create_ummg(configuration: config.Config, granule: Granule) -> Granule:
     # }
     metadata_details = {}
     for data_file in granule.data_filenames:
-        metadata_details[data_file] = granule.data_reader(data_file, configuration)
+        metadata_details[data_file] = granule.data_reader(
+            data_file, granule.premet_filename, configuration
+        )
 
     # Collapse information about (possibly) multiple files into a granule summary.
     summary = metadata_summary(metadata_details)