+++ tests/test_metgen.py
@@ -17,6 +17,26 @@ from nsidc.metgen import config, constants, metgen
 
 
 @pytest.fixture
+def test_config():
+    return config.Config(
+        environment="uat",
+        staging_bucket_name="cloud_bucket",
+        data_dir="foo",
+        auth_id="nsidc-0000",
+        version=1,
+        provider="blah",
+        local_output_dir="output",
+        ummg_dir="ummg",
+        kinesis_stream_name="fake_stream",
+        write_cnm_file=True,
+        overwrite_ummg=True,
+        checksum_type="sha",
+        number=3,
+        dry_run=False,
+    )
+
+
+@pytest.fixture
 def multi_file_granule():
     return {
         "first_id": {
@@ -128,113 +148,180 @@ def test_granule_name_from_regex(regex):
 
 
 @pytest.mark.parametrize(
-    "granuleid,data_files,browse_files,expected",
+    "granuleid,data_files,browse_files,premet_files,expected",
     [
         (
             "aaa_gid1_bbb",
             ["aaa_gid1_bbb.nc"],
             [],
-            ("aaa_gid1_bbb.nc", {"aaa_gid1_bbb.nc"}, set()),
+            [],
+            ("aaa_gid1_bbb.nc", {"aaa_gid1_bbb.nc"}, set(), ""),
         ),
         (
             "aaa_gid1_bbb",
             ["aaa_gid1_bbb.nc"],
             ["aaa_gid1_browse_bbb.png"],
-            ("aaa_gid1_bbb.nc", {"aaa_gid1_bbb.nc"}, set()),
+            [],
+            ("aaa_gid1_bbb.nc", {"aaa_gid1_bbb.nc"}, set(), ""),
+        ),
+        (
+            "aaa_gid1_bbb",
+            ["aaa_gid1_bbb.nc"],
+            ["aaa_gid1_browse_bbb.png"],
+            ["aaa_gid1_bbb.nc.premet"],
+            (
+                "aaa_gid1_bbb.nc",
+                {"aaa_gid1_bbb.nc"},
+                set(),
+                "aaa_gid1_bbb.nc.premet",
+            ),
         ),
         (
             "aaa_gid1_bbb",
             ["aaa_gid1_bbb.nc"],
             ["aaa_gid1_bbb_browse.png"],
-            ("aaa_gid1_bbb.nc", {"aaa_gid1_bbb.nc"}, {"aaa_gid1_bbb_browse.png"}),
+            [],
+            ("aaa_gid1_bbb.nc", {"aaa_gid1_bbb.nc"}, {"aaa_gid1_bbb_browse.png"}, ""),
+        ),
+        (
+            "aaa_gid1_bbb",
+            ["aaa_gid1_bbb.nc"],
+            ["aaa_gid1_bbb_browse.png"],
+            ["aaa_gid1_bbb.nc.premet"],
+            (
+                "aaa_gid1_bbb.nc",
+                {"aaa_gid1_bbb.nc"},
+                {"aaa_gid1_bbb_browse.png"},
+                "aaa_gid1_bbb.nc.premet",
+            ),
+        ),
+        (
+            "aaa_gid1_bbb",
+            ["aaa_gid1_bbb.nc"],
+            ["aaa_gid1_bbb_browse.png"],
+            ["ccc_gid1_ddd.nc.premet"],
+            ("aaa_gid1_bbb.nc", {"aaa_gid1_bbb.nc"}, {"aaa_gid1_bbb_browse.png"}, ""),
         ),
         (
             "aaa_gid1_bbb",
             ["aaa_gid1_bbb.nc", "aaa_gid1_bbb.tif"],
             ["aaa_gid1_bbb_browse.png"],
+            [],
             (
                 "aaa_gid1_bbb",
                 {"aaa_gid1_bbb.nc", "aaa_gid1_bbb.tif"},
                 {"aaa_gid1_bbb_browse.png"},
+                "",
+            ),
+        ),
+        (
+            "aaa_gid1_bbb",
+            ["aaa_gid1_bbb.nc", "aaa_gid1_bbb.tif"],
+            ["aaa_gid1_bbb_browse.png"],
+            ["aaa_gid1_bbb.premet"],
+            (
+                "aaa_gid1_bbb",
+                {"aaa_gid1_bbb.nc", "aaa_gid1_bbb.tif"},
+                {"aaa_gid1_bbb_browse.png"},
+                "aaa_gid1_bbb.premet",
             ),
         ),
         (
             "aaa_gid1_bbb",
             ["aaa_gid1_bbb.nc", "aaa_gid1_bbb.tif"],
             ["aaa_gid1_bbb_browse.png", "aaa_gid1_browse_bbb.tif"],
+            [],
             (
                 "aaa_gid1_bbb",
                 {"aaa_gid1_bbb.nc", "aaa_gid1_bbb.tif"},
                 {"aaa_gid1_bbb_browse.png"},
+                "",
             ),
         ),
     ],
 )
-def test_granule_tuple_from_filenames(granuleid, data_files, browse_files, expected):
+def test_granule_tuple_from_filenames(
+    granuleid, data_files, browse_files, premet_files, expected
+):
     granule = metgen.granule_tuple(
         granuleid,
         f"({granuleid})",
         "browse",
         [Path(p) for p in data_files + browse_files],
+        [Path(p) for p in premet_files],
     )
     assert granule == expected
 
 
 @pytest.mark.parametrize(
-    "granuleid,data_files,browse_files,expected",
+    "granuleid,data_files,browse_files,premet_files,expected",
     [
         (
             "gid1",
             ["aaa_gid1_bbb.nc", "aaa_gid1_bbb.tif"],
             [],
-            ("aaa_gid1_bbb", {"aaa_gid1_bbb.nc", "aaa_gid1_bbb.tif"}, set()),
+            [],
+            ("aaa_gid1_bbb", {"aaa_gid1_bbb.nc", "aaa_gid1_bbb.tif"}, set(), ""),
         ),
         (
             "gid1",
             ["aaa_gid1_bbb.nc", "aaa_gid1_bbb.tif"],
             ["aaa_gid1_browse_bbb.png"],
+            ["aaa_gid1_bbb.premet"],
             (
                 "aaa_gid1_bbb",
                 {"aaa_gid1_bbb.nc", "aaa_gid1_bbb.tif"},
                 {"aaa_gid1_browse_bbb.png"},
+                "aaa_gid1_bbb.premet",
             ),
         ),
         (
             "gid1",
             ["aaa_gid1_xx_bbb.nc", "aaa_gid1_bbb.tif"],
             ["aaa_gid1_browse_bbb.png"],
+            ["aaa_gid1_xx_bbb.premet"],
             (
                 "aaa_gid1_bbb",
                 {"aaa_gid1_xx_bbb.nc", "aaa_gid1_bbb.tif"},
                 {"aaa_gid1_browse_bbb.png"},
+                "aaa_gid1_xx_bbb.premet",
             ),
         ),
         (
             "gid1",
             ["aaa_gid1_zz_bbb.nc", "aaa_gid1_xx_bbb.tif"],
             ["aaa_gid1_browse_zz_bbb.png", "aaa_gid1_browse_yy_bbb.tif"],
+            [],
             (
                 "aaa_gid1_bbb",
                 {"aaa_gid1_zz_bbb.nc", "aaa_gid1_xx_bbb.tif"},
                 {"aaa_gid1_browse_zz_bbb.png", "aaa_gid1_browse_yy_bbb.tif"},
+                "",
             ),
         ),
     ],
 )
-def test_granule_tuple_from_regex(granuleid, data_files, browse_files, expected, regex):
+def test_granule_tuple_from_regex(
+    granuleid, data_files, browse_files, premet_files, expected, regex
+):
     granule = metgen.granule_tuple(
         granuleid,
         regex,
         "browse",
         [Path(p) for p in data_files + browse_files],
+        [Path(p) for p in premet_files],
     )
     assert granule == expected
+
+
+def test_with_no_premet_dir(test_config):
+    assert metgen.premet_files(test_config) is None
 
 
 @patch("nsidc.metgen.metgen.s3_object_path", return_value="/some/path")
 @patch("nsidc.metgen.aws.stage_file", return_value=True)
 @patch("builtins.open", new_callable=mock_open, read_data="data")
-def test_stage_files(m1, m2, m3):
+def test_stage_files(m1, m2, m3, test_config):
     granule = metgen.Granule(
         "foo",
         metgen.Collection("ABCD", 2),
@@ -243,23 +330,7 @@ def test_stage_files(m1, m2, m3):
         browse_filenames={"browse1", "browse2", "browse3"},
         ummg_filename="foo_ummg",
     )
-    configuration = config.Config(
-        "uat",
-        staging_bucket_name="cloud_bucket",
-        data_dir="foo",
-        auth_id="nsidc-0000",
-        version=1,
-        provider="blah",
-        local_output_dir="output",
-        ummg_dir="ummg",
-        kinesis_stream_name="fake_stream",
-        write_cnm_file=True,
-        overwrite_ummg=True,
-        checksum_type="sha",
-        number=3,
-        dry_run=False,
-    )
-    assert metgen.stage_files(configuration, granule)
+    assert metgen.stage_files(test_config, granule)
 
 
 def test_returns_datetime_range():