+++ tests/test_reader_lookup.py
@@ -1,5 +1,5 @@
 import pytest
-from nsidc.metgen import netcdf_reader
+from nsidc.metgen import constants, netcdf_reader
 from nsidc.metgen.readers import csv, snowex_csv
 from nsidc.metgen.readers.registry import lookup
 
@@ -7,9 +7,9 @@ from nsidc.metgen.readers.registry import lookup
 @pytest.mark.parametrize(
     "collection,extension,expected",
     [
-        ("NSIDC-0081DUCk", ".nc", netcdf_reader.extract_metadata),
-        ("SNEX23_SSADUCk", ".csv", snowex_csv.extract_metadata),
-        ("IRWIS2DUCk", ".csv", csv.extract_metadata),
+        ("NSIDC-0081DUCk", constants.NETCDF_SUFFIX, netcdf_reader.extract_metadata),
+        ("SNEX23_SSADUCk", constants.CSV_SUFFIX, snowex_csv.extract_metadata),
+        ("IRWIS2DUCk", constants.CSV_SUFFIX, csv.extract_metadata),
     ],
 )
 def test_reader(collection, extension, expected):